# Conflicts:
#	.gitignore
#	PolicyManager/doc/Install-Config-Test.md

+++ dependent_pkg/download.sh
@@ -14,29 +14,17 @@ get_or_update_repo() {
 cd $(dirname `which $0`)  # cd to script dir
 
 # --- json2html ---
-repodir='json2html'   # TODO: use wildcard to remove version
+repodir='json2html'
 repourl='https://github.com/rhoerbe/json2html.git'
 get_or_update_repo
 
 
 # --- pyjnius ---
-if [[ "$ostype" == 'darwin' ]]; then
-    repodir='benson-basis/pyjnius'
-    repourl='https://github.com/benson-basis/pyjnius.git'
-else
-    repodir='kivy/pyjnius'
-    repourl='https://github.com/kivy/pyjnius.git'
-fi
+repodir='kivy/pyjnius'
+repourl='https://github.com/kivy/pyjnius.git'
 get_or_update_repo
-ln -s $repodir pyjnius
+[ -e "$repodir/pyjnius" ] || ln -s $repodir pyjnius
 
-# --- ordereddict (for seom versions of json2html) ---
-if [ ! -e ordereddict-1.1 ] ; then
-    echo "downloading ordereddict-1.1"
-	curl -O https://pypi.python.org/packages/source/o/ordereddict/ordereddict-1.1.tar.gz
-	tar -xzf ordereddict-1.1.tar.gz
-	rm ordereddict-1.1.tar.gz
-fi
